Organize `Cardano.Ledger.Alonzo.Core` exports of type classes:
* Move core type class definitions to modules where instances are defined.
* Remove unnecessary re-exports
* Remove derpecated functions
